Batch changes to project
Full change to each individual locality with dynamic data correlating to them. Each location is comprised of photographs, locality info & history, pricing and lat and long values for the google map marker. Also, comments added or deleted by the user now redirect him back to the success message so there is no need for scrolling. TODO: Add pictures to the localities grid, add styling to the main page

+++ database/migrations/2016_08_09_145412_create_localities_table.php
@@ -22,8 +22,8 @@ class CreateLocalitiesTable extends Migration
             $table->integer('price')->unsigned()->default(0);
             $table->integer('price_group')->unsigned()->default(0);
             $table->integer('price_child')->unsigned()->default(0);
-            $table->float('lat');
-            $table->float('long');
+            $table->double('lat');
+            $table->double('long');
             $table->timestamps();
         });
     }
